Rework google sign in button to be custom
